fix: extra stat on filehandle
