+++ ext/fs/std_fs.rs
@@ -862,7 +862,7 @@ fn stat(path: &Path) -> FsResult<FsStat> {
   let file = opts.open(path)?;
   let metadata = file.metadata()?;
   let mut fsstat = FsStat::from_std(metadata);
-  stat_extra(&file, &mut fsstat)?;
+  deno_io::stat_extra(&file, &mut fsstat)?;
   Ok(fsstat)
 }
 
@@ -885,130 +885,8 @@ fn lstat(path: &Path) -> FsResult<FsStat> {
   let file = opts.open(path)?;
   let metadata = file.metadata()?;
   let mut fsstat = FsStat::from_std(metadata);
-  stat_extra(&file, &mut fsstat)?;
+  deno_io::stat_extra(&file, &mut fsstat)?;
   Ok(fsstat)
-}
-
-#[cfg(windows)]
-fn stat_extra(file: &std::fs::File, fsstat: &mut FsStat) -> FsResult<()> {
-  use std::os::windows::io::AsRawHandle;
-
-  unsafe fn get_dev(
-    handle: winapi::shared::ntdef::HANDLE,
-  ) -> std::io::Result<u64> {
-    use winapi::shared::minwindef::FALSE;
-    use winapi::um::fileapi::BY_HANDLE_FILE_INFORMATION;
-    use winapi::um::fileapi::GetFileInformationByHandle;
-
-    // SAFETY: winapi calls
-    unsafe {
-      let info = {
-        let mut info =
-          std::mem::MaybeUninit::<BY_HANDLE_FILE_INFORMATION>::zeroed();
-        if GetFileInformationByHandle(handle, info.as_mut_ptr()) == FALSE {
-          return Err(std::io::Error::last_os_error());
-        }
-
-        info.assume_init()
-      };
-
-      Ok(info.dwVolumeSerialNumber as u64)
-    }
-  }
-
-  const WINDOWS_TICK: i64 = 10_000; // 100-nanosecond intervals in a millisecond
-  const SEC_TO_UNIX_EPOCH: i64 = 11_644_473_600; // Seconds between Windows epoch and Unix epoch
-
-  fn windows_time_to_unix_time_msec(windows_time: &i64) -> i64 {
-    let milliseconds_since_windows_epoch = windows_time / WINDOWS_TICK;
-    milliseconds_since_windows_epoch - SEC_TO_UNIX_EPOCH * 1000
-  }
-
-  use windows_sys::Wdk::Storage::FileSystem::FILE_ALL_INFORMATION;
-  use windows_sys::Win32::Foundation::NTSTATUS;
-
-  unsafe fn query_file_information(
-    handle: winapi::shared::ntdef::HANDLE,
-  ) -> Result<FILE_ALL_INFORMATION, NTSTATUS> {
-    use windows_sys::Wdk::Storage::FileSystem::NtQueryInformationFile;
-    use windows_sys::Win32::Foundation::ERROR_MORE_DATA;
-    use windows_sys::Win32::Foundation::RtlNtStatusToDosError;
-    use windows_sys::Win32::System::IO::IO_STATUS_BLOCK;
-
-    // SAFETY: winapi calls
-    unsafe {
-      let mut info = std::mem::MaybeUninit::<FILE_ALL_INFORMATION>::zeroed();
-      let mut io_status_block =
-        std::mem::MaybeUninit::<IO_STATUS_BLOCK>::zeroed();
-      let status = NtQueryInformationFile(
-        handle as _,
-        io_status_block.as_mut_ptr(),
-        info.as_mut_ptr() as *mut _,
-        std::mem::size_of::<FILE_ALL_INFORMATION>() as _,
-        18, /* FileAllInformation */
-      );
-
-      if status < 0 {
-        let converted_status = RtlNtStatusToDosError(status);
-
-        // If error more data is returned, then it means that the buffer is too small to get full filename information
-        // to have that we should retry. However, since we only use BasicInformation and StandardInformation, it is fine to ignore it
-        // since struct is populated with other data anyway.
-        // https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/ntifs/nf-ntifs-ntqueryinformationfile#remarksdd
-        if converted_status != ERROR_MORE_DATA {
-          return Err(converted_status as NTSTATUS);
-        }
-      }
-
-      Ok(info.assume_init())
-    }
-  }
-
-  // SAFETY: winapi calls
-  unsafe {
-    let file_handle = file.as_raw_handle();
-
-    fsstat.dev = get_dev(file_handle)?;
-
-    if let Ok(file_info) = query_file_information(file_handle) {
-      fsstat.ctime = Some(windows_time_to_unix_time_msec(
-        &file_info.BasicInformation.ChangeTime,
-      ) as u64);
-
-      if file_info.BasicInformation.FileAttributes
-        & winapi::um::winnt::FILE_ATTRIBUTE_REPARSE_POINT
-        != 0
-      {
-        fsstat.is_symlink = true;
-      }
-
-      if file_info.BasicInformation.FileAttributes
-        & winapi::um::winnt::FILE_ATTRIBUTE_DIRECTORY
-        != 0
-      {
-        fsstat.mode |= libc::S_IFDIR as u32;
-        fsstat.size = 0;
-      } else {
-        fsstat.mode |= libc::S_IFREG as u32;
-        fsstat.size = file_info.StandardInformation.EndOfFile as u64;
-      }
-
-      if file_info.BasicInformation.FileAttributes
-        & winapi::um::winnt::FILE_ATTRIBUTE_READONLY
-        != 0
-      {
-        fsstat.mode |=
-          (libc::S_IREAD | (libc::S_IREAD >> 3) | (libc::S_IREAD >> 6)) as u32;
-      } else {
-        fsstat.mode |= ((libc::S_IREAD | libc::S_IWRITE)
-          | ((libc::S_IREAD | libc::S_IWRITE) >> 3)
-          | ((libc::S_IREAD | libc::S_IWRITE) >> 6))
-          as u32;
-      }
-    }
-
-    Ok(())
-  }
 }
 
 fn exists(path: &Path) -> bool {
